+++ src/app/styles/constants.tsx
@@ -11,6 +11,13 @@ const constants = {
     cream: '#f8faf3',
     gunmetal: '#2b303a',
   },
+  fontWeights: {
+    light: 300,
+    regular: 400,
+    medium: 500,
+    semiBold: 600,
+    bold: 700,
+  },
 };
 
 export default constants;